feat(Command) : 결과를 보고 싶은 사람에 대한 입력을 Command 클래스로 다루도록 변경
